refactor(source): Pull in read_packages
This is tied to the `Source` and only used there.
